scraper-service/requirements.txt
requests==2.32.5
kafka-python==2.3.0
redis==7.3.0
python-dotenv==1.2.2
